# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.RequestAuth:
    properties:
      email:
        type: string
      password:
        maxLength: 100
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
...
  common.ResponseError:
    properties:
      error:
        type: string
    type: object
...
  common.ResponseNormal:
    properties:
      data: {}
      email:
        type: string
      message:
        type: string
      paging: {}
      result:
        type: boolean
    type: object
...
  models.Account:
    properties:
      account_status:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      is_verified:
        type: boolean
      password:
        maxLength: 100
        minLength: 8
        type: string
      role:
        type: string
    required:
    - email
    - password
    type: object
  models.AccountCreate:
    properties:
      email:
        type: string
      is_verified:
        type: boolean
      password:
        maxLength: 100
        minLength: 8
        type: string
      role:
        enum:
        - admin
        - user
        type: string
    required:
    - email
    - password
    type: object
...
host: 127.0.0.1:9000
info:
  contact: {}
  description: List APIs of Healthy Management Service
  termsOfService: http://swagger.io/terms/
  title: Healthy Service API Document
  version: "1.0"
paths:
...
  /admin/experts/accounts:
    post:
      consumes:
      - application/json
      description: Create a new expert account with email and password
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Account information
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.AccountCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Create expert account successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  $ref: '#/definitions/models.Account'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: Invalid token
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: You do not have permission to access this resource
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Create a new expert account
      tags:
      - Expert
  /admin/experts/accounts/{id}/lock:
    patch:
      consumes:
      - application/json
      description: Lock expert account by expert ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Expert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Expert account locked successfully
          schema:
            $ref: '#/definitions/common.ResponseNormal'
        "400":
          description: Expert ID is required
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: Invalid token
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: You do not have permission to access this resource
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Lock expert account
      tags:
      - Expert
  /admin/experts/accounts/{id}/unlock:
    patch:
      consumes:
      - application/json
      description: Unlock expert account by expert ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Expert ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Expert account unlocked successfully
          schema:
            $ref: '#/definitions/common.ResponseNormal'
        "400":
          description: Expert ID is required
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: Invalid token
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: You do not have permission to access this resource
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Unlock expert account
      tags:
      - Expert
  /admin/experts/accounts/reset-password:
    post:
      consumes:
      - application/json
      description: Reset expert password with email and new password
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Reset password request
        in: body
        name: resetPasswordRequest
        required: true
        schema:
          $ref: '#/definitions/common.RequestAuth'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            $ref: '#/definitions/common.ResponseNormal'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.ResponseError'
        "401":
          description: Invalid token
          schema:
            $ref: '#/definitions/common.ResponseError'
        "403":
          description: You do not have permission to access this resource
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Reset expert password
      tags:
      - Expert
  /admin/users:
    get:
      consumes:
      - application/json
      description: Get a list of users with pagination
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Page number (default is 1)
        in: query
        name: page
        type: integer
      - description: Number of users per page (default is 10)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of users
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Account'
                  type: array
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Get list of users
      tags:
      - User
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Account information
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.AccountCreate'
      produces:
      - application/json
      responses:
        "201":
          description: Account created successfully
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  $ref: '#/definitions/models.Account'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Create a new user account
      tags:
      - User
  /admin/users/{id}:
    get:
      consumes:
      - application/json
      description: Get user details by user ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseNormal'
            - properties:
                data:
                  $ref: '#/definitions/models.Account'
              type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/common.ResponseError'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Get user by ID
      tags:
      - User
  /admin/users/{id}/lock:
    patch:
      consumes:
      - application/json
      description: Lock user account by user ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User account locked successfully
          schema:
            $ref: '#/definitions/common.ResponseNormal'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Lock user account
      tags:
      - User
  /admin/users/{id}/unlock:
    patch:
      consumes:
      - application/json
      description: Unlock user account by user ID
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User account unlocked successfully
          schema:
            $ref: '#/definitions/common.ResponseNormal'
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Unlock user account
      tags:
      - User
  /admin/users/reset-password:
    post:
      consumes:
      - application/json
      description: Reset user password with email and new password
      parameters:
      - description: Bearer Token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Reset password request
        in: body
        name: resetPasswordRequest
        required: true
        schema:
          $ref: '#/definitions/common.RequestAuth'
      produces:
      - application/json
      responses:
        "200":
          description: Password reset successfully
          schema:
            $ref: '#/definitions/common.ResponseNormal'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/common.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/common.ResponseError'
      summary: Reset user password
      tags:
      - User
...
schemes:
- http
- https
swagger: "2.0"